requirements.txt
python-whois
prettytable